Add client program and make client and server exit gracefully for #6
Make the server exit gracefully on a keyboard interrupt (Ctrl+C) and add
a client program that takes the hostname of the server to connect to,
and optionally the port (otherwise uses 2222). Client sends user input
to the server as newline delimited messages. Client exits gracefully if
requested by the user or if the server socket is closed.